Add files via upload
Here we are using MNIST dataset of 60,000 images in total and we are only loading those images for our project. The dataset is then further divided into two sets; Training and Testing which is further processed through Feature Engineering that makes it easier for our KNN classifier to label the images. At the final stage, we are provided with an output of images, each with an accurate label. The validation accuracy shown is 0.95525 and the testing accuracy is 0.9509.